Move test_macros.rs -> macros.rs
